Fix file handling issue reported by Mike Keith:
It crashes in getecliptic.f when $TEMPO is on a read-only filesystem (as
happens to be the case for our NFS-mounted version of tempo on some systems).
This only appears on some computers, and is due to a subtle mismatch between
the way fortran and posix file handing works.

In any case, the fix is very simple, just adding "action='read', status='old'" to the open statement for ecliptic.dat. (this prevents it from trying to open the file with "Create if doesn't exist" flags, which causes an error on some read-only filesystems even if the file does exist!)